Added CRITICAL_UP event check in UMM.
