+++ cluster/src/dunit/scala/org/apache/spark/memory/SnappyUnifiedMemoryManagerDUnitTest.scala
@@ -29,22 +29,22 @@ import SnappyUnifiedMemoryManagerDUnitTest._
 
 case class DummyData(col1: Int, col2: Int, col3: Int)
 
-class WaitAssert(val error: Int, clazz : Class[_]){
+class WaitAssert(val error: Int, clazz: Class[_]) {
 
   var value1 = 0L
   var value2 = 0L
   var excString = ""
 
-  def assertStorageUsed(vm1: VM, vm2: VM, ignoreByteCount :Int = 0): Boolean = {
+  def assertStorageUsed(vm1: VM, vm2: VM, ignoreByteCount: Int = 0): Boolean = {
     value1 = vm1.invoke(clazz, "getStorageMemory").asInstanceOf[Long]
     value2 = vm2.invoke(clazz, "getStorageMemory").asInstanceOf[Long]
     println(s"vm1_memoryUsed $value1 vm2_memoryUsed $value2")
     excString = s"failed $value1 & $value2 are not within permissable limit \n"
 
     if (value1 == value2) return true
-    if(value1 < value2 ){
+    if (value1 < value2) {
       value1 += ignoreByteCount
-    }else{
+    } else {
       value2 -= ignoreByteCount
     }
     if (Math.abs(value1 - value2) < ((value2 * error) / 100)) return true else false
@@ -61,6 +61,7 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
   val memoryMode = MemoryMode.ON_HEAP
 
   bootProps.setProperty(io.snappydata.Property.CachedBatchSize.name, "500")
+  bootProps.setProperty("critical-heap-percentage", "90")
 
   def newContext(): SnappyContext = {
     val snc = SnappyContext(sc).newSession()
@@ -69,11 +70,12 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
 
   }
 
-  def resetMemoryManagers(): Unit ={
+  def resetMemoryManagers(): Unit = {
     vm0.invoke(getClass, "resetStorageMemory")
     vm1.invoke(getClass, "resetStorageMemory")
     vm2.invoke(getClass, "resetStorageMemory")
   }
+
   override def tearDown2(): Unit = {
     resetMemoryManagers
     super.tearDown2()
@@ -82,7 +84,8 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
   //Approximate because we include hash map size also, which can vary across VMs
   def assertApproximate(value1: Long, value2: Long, error: Int = 5): Unit = {
     if (value1 == value2) return
-    if(Math.abs(value1 - value2) > ((value2 * error)/ 100)){// Error target should be 1
+    if (Math.abs(value1 - value2) > ((value2 * error) / 100)) {
+      // Error target should be 1
       throw new java.lang.AssertionError(s"assertion " +
           s"failed $value1 & $value2 are not within permissable limit")
     }
@@ -90,10 +93,10 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
 
   def assertForWait(value1: Long, value2: Long, error: Int = 5): Boolean = {
     if (value1 == value2) return true
-    if(Math.abs(value1 - value2) < ((value2 * error)/ 100)) return true else false
+    if (Math.abs(value1 - value2) < ((value2 * error) / 100)) return true else false
   }
 
-  def _testMemoryUsedInReplication(): Unit = {
+  def testMemoryUsedInReplication(): Unit = {
     val snc = newContext()
     val data = for (i <- 1 to 500) yield (Seq(i, (i + 1), (i + 2)))
     val rdd = snc.sparkContext.parallelize(data.toSeq, data.length).map(s =>
@@ -104,7 +107,43 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
     val vm1_memoryUsed = vm1.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
     val vm2_memoryUsed = vm2.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
     println(s"vm1_memoryUsed $vm1_memoryUsed vm2_memoryUsed $vm2_memoryUsed")
-    assertApproximate(vm1_memoryUsed , vm2_memoryUsed)
+    assertApproximate(vm1_memoryUsed, vm2_memoryUsed)
+  }
+
+  def testMemoryUsedInBucketRegions_RowTables(): Unit = {
+    val snc = newContext()
+    val data = for (i <- 1 to 500) yield (Seq(i, (i + 1), (i + 2)))
+    val rdd = snc.sparkContext.parallelize(data.toSeq, data.length).map(s =>
+      DummyData(s(0), s(1), s(2)))
+    val dataDF = snc.createDataFrame(rdd)
+    val options = "OPTIONS (BUCKETS '113', PARTITION_BY 'Col1', REDUNDANCY '2')"
+    snc.sql("CREATE TABLE " + rr_table + " (Col1 INT, Col2 INT, Col3 INT) " + " USING row " +
+        options
+    )
+    dataDF.write.insertInto(rr_table)
+
+    val vm1_memoryUsed = vm1.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
+    val vm2_memoryUsed = vm2.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
+    println(s"vm1_memoryUsed $vm1_memoryUsed vm2_memoryUsed $vm2_memoryUsed")
+    assertApproximate(vm1_memoryUsed, vm2_memoryUsed)
+  }
+
+  def testMemoryUsedInBucketRegions_ColumntTables(): Unit = {
+    val snc = newContext()
+    val data = for (i <- 1 to 500) yield (Seq(i, (i + 1), (i + 2)))
+    val rdd = snc.sparkContext.parallelize(data.toSeq, data.length).map(s =>
+      DummyData(s(0), s(1), s(2)))
+    val dataDF = snc.createDataFrame(rdd)
+    val options = "OPTIONS (BUCKETS '113', PARTITION_BY 'Col1', REDUNDANCY '2')"
+    snc.sql("CREATE TABLE " + col_table + " (Col1 INT, Col2 INT, Col3 INT) " + " USING column " +
+        options
+    )
+    dataDF.write.insertInto(col_table)
+
+    val vm1_memoryUsed = vm1.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
+    val vm2_memoryUsed = vm2.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
+    println(s"vm1_memoryUsed $vm1_memoryUsed vm2_memoryUsed $vm2_memoryUsed")
+    assertApproximate(vm1_memoryUsed, vm2_memoryUsed)
   }
 
   /**
@@ -116,6 +155,7 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
     var props = bootProps.clone().asInstanceOf[java.util.Properties]
     val port = ClusterManagerTestBase.locPort
     props.setProperty("eviction-heap-percentage", "80")
+
     def restartServer(props: Properties): SerializableRunnable = new SerializableRunnable() {
       override def run(): Unit = ClusterManagerTestBase.startSnappyServer(port, props)
     }
@@ -135,7 +175,7 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
 
     val waitAssert = new WaitAssert(2, getClass)
     //Setting ignore bytecount as VM doing GII does have a valid value, hence key is kept as null
-    // This decreases the size of entry overhead
+    // This decreases the size of entry overhead. @TODO find out why only column table needs this ?
     ClusterManagerTestBase.waitForCriterion(waitAssert.assertStorageUsed(vm1, vm2, 16000),
       waitAssert.exceptionString(),
       20000, 5000, true)
@@ -146,10 +186,11 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
     * It checks memory usage with reference to the node which was alive at the time
     * of GII.
     */
-  def testMemoryUsedInReplicationWithGII(): Unit = {
+  def testMemoryUsedInReplicatedTableWithGII(): Unit = {
     var props = bootProps.clone().asInstanceOf[java.util.Properties]
     val port = ClusterManagerTestBase.locPort
     props.setProperty("eviction-heap-percentage", "80")
+
     def restartServer(props: Properties): SerializableRunnable = new SerializableRunnable() {
       override def run(): Unit = ClusterManagerTestBase.startSnappyServer(port, props)
     }
@@ -175,10 +216,9 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
     * It checks memory usage with reference to the node which was alive at the time
     * of GII.
     */
-  def _testMemoryUsedInRowPartitionedTableWithGII(): Unit = {
-    var props = bootProps.clone().asInstanceOf[java.util.Properties]
+  def testMemoryUsedInRowPartitionedTableWithGII(): Unit = {
+    val props = bootProps.clone().asInstanceOf[java.util.Properties]
     val port = ClusterManagerTestBase.locPort
-    props.setProperty("eviction-heap-percentage", "80")
 
     def restartServer(props: Properties): SerializableRunnable = new SerializableRunnable() {
       override def run(): Unit = ClusterManagerTestBase.startSnappyServer(port, props)
@@ -209,10 +249,9 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
     * It checks memory usage with reference to the node which was alive at the time
     * of GII. At the same time we fire deletes on the region.
     */
-  def _testMemoryUsedInReplicationParTableGIIWithDeletes(): Unit = {
-    var props = bootProps.clone().asInstanceOf[java.util.Properties]
+  def testMemoryUsedInReplicationParTableGIIWithDeletes(): Unit = {
+    val props = bootProps.clone().asInstanceOf[java.util.Properties]
     val port = ClusterManagerTestBase.locPort
-    props.setProperty("eviction-heap-percentage", "80")
 
     def restartServer(props: Properties): SerializableRunnable = new SerializableRunnable() {
       override def run(): Unit = ClusterManagerTestBase.startSnappyServer(port, props)
@@ -232,46 +271,62 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
     dataDF.write.insertInto(rr_table)
 
     val otherExecutorThread = new Thread(new Runnable {
+
       def run() {
-        (1 to 100).map(i=> snc.delete(rr_table, s"col1=$i"))
+        (1 to 100).map(i => snc.delete(rr_table, s"col1=$i"))
       }
     })
     otherExecutorThread.start()
 
     vm1.invoke(restartServer(props))
-    val vm1_memoryUsed = vm1.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
-    val vm2_memoryUsed = vm2.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
-    println(s"vm1_memoryUsed $vm1_memoryUsed vm2_memoryUsed $vm2_memoryUsed")
 
-    ClusterManagerTestBase.waitForCriterion(assertForWait(vm1.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
-      , vm2.invoke(getClass, "getStorageMemory").asInstanceOf[Long], 2),
-      s"failed $vm1_memoryUsed & $vm2_memoryUsed are not within permissable limit \n",
-      20000, 5000, true)
+    val waitAssert = new WaitAssert(2, getClass)
+    //The delete operation takes time to propagate
+    ClusterManagerTestBase.waitForCriterion(waitAssert.assertStorageUsed(vm1, vm2),
+      waitAssert.exceptionString(),
+      30000, 5000, true)
   }
 
-  def _testMemoryUsedInBucketRegions(): Unit = {
+
+  def testMemoryAfterRecovery_ColumnTable(): Unit = {
+    val props = bootProps.clone().asInstanceOf[java.util.Properties]
+    val port = ClusterManagerTestBase.locPort
+
+    def restartServer(props: Properties): SerializableRunnable = new SerializableRunnable() {
+      override def run(): Unit = ClusterManagerTestBase.startSnappyServer(port, props)
+    }
+
     val snc = newContext()
+
     val data = for (i <- 1 to 500) yield (Seq(i, (i + 1), (i + 2)))
     val rdd = snc.sparkContext.parallelize(data.toSeq, data.length).map(s =>
       DummyData(s(0), s(1), s(2)))
     val dataDF = snc.createDataFrame(rdd)
-    val options = "OPTIONS (BUCKETS '113', PARTITION_BY 'Col1', REDUNDANCY '2')"
-    snc.sql("CREATE TABLE " + rr_table + " (Col1 INT, Col2 INT, Col3 INT) " + " USING row " +
+    val options = "OPTIONS (BUCKETS '113', PARTITION_BY 'Col1', PERSISTENT 'SYNCHRONOUS', REDUNDANCY '2')"
+    snc.sql("CREATE TABLE " + col_table + " (Col1 INT, Col2 INT, Col3 INT) " + " USING column " +
         options
     )
-    dataDF.write.insertInto(rr_table)
-
-    val vm1_memoryUsed = vm1.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
+    dataDF.write.insertInto(col_table)
+    var vm1_memoryUsed = vm1.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
     val vm2_memoryUsed = vm2.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
     println(s"vm1_memoryUsed $vm1_memoryUsed vm2_memoryUsed $vm2_memoryUsed")
-    assertApproximate(vm1_memoryUsed , vm2_memoryUsed)
+    assertApproximate(vm1_memoryUsed, vm2_memoryUsed)
+    vm1.invoke(classOf[ClusterManagerTestBase], "stopAny")
+
+    vm1.invoke(restartServer(props))
+
+    val waitAssert = new WaitAssert(2, getClass)
+    //The delete operation takes time to propagate
+    ClusterManagerTestBase.waitForCriterion(waitAssert.assertStorageUsed(vm1, vm2),
+      waitAssert.exceptionString(),
+      30000, 5000, true)
+
   }
 
-  def _testMemoryAfterRecovery(): Unit = {
-    var props = bootProps.clone().asInstanceOf[java.util.Properties]
-    val port = ClusterManagerTestBase.locPort
 
-    props.setProperty("eviction-heap-percentage", "80")
+  def testMemoryAfterRecovery_RowTable(): Unit = {
+    val props = bootProps.clone().asInstanceOf[java.util.Properties]
+    val port = ClusterManagerTestBase.locPort
 
     def restartServer(props: Properties): SerializableRunnable = new SerializableRunnable() {
       override def run(): Unit = ClusterManagerTestBase.startSnappyServer(port, props)
@@ -291,20 +346,22 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
     var vm1_memoryUsed = vm1.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
     val vm2_memoryUsed = vm2.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
     println(s"vm1_memoryUsed $vm1_memoryUsed vm2_memoryUsed $vm2_memoryUsed")
-    assertApproximate(vm1_memoryUsed , vm2_memoryUsed)
+    assertApproximate(vm1_memoryUsed, vm2_memoryUsed)
     vm1.invoke(classOf[ClusterManagerTestBase], "stopAny")
-    props = bootProps
 
     vm1.invoke(restartServer(props))
-    val post_restart_vm1_memoryUsed = vm1.invoke(getClass, "getStorageMemory").asInstanceOf[Long]
-    println(s"vm1_memoryUsed $vm1_memoryUsed post_restart_vm1_memoryUsed $post_restart_vm1_memoryUsed")
-    assertApproximate(vm1_memoryUsed , post_restart_vm1_memoryUsed)
-
+    
+    val waitAssert = new WaitAssert(2, getClass)
+    //The delete operation takes time to propagate
+    ClusterManagerTestBase.waitForCriterion(waitAssert.assertStorageUsed(vm1, vm2),
+      waitAssert.exceptionString(),
+      30000, 5000, true)
   }
 }
 
 object SnappyUnifiedMemoryManagerDUnitTest {
   private def sc = SnappyContext.globalSparkContext
+
   val memoryMode = MemoryMode.ON_HEAP
 
   def resetStorageMemory() =
@@ -320,7 +377,7 @@ object SnappyUnifiedMemoryManagerDUnitTest {
     }.collect()
   }
 
-  def failAllExecutors : Unit = {
+  def failAllExecutors: Unit = {
     try {
       failTheExecutors
     } catch {
